Add files via upload
This is a small tasklist app made with vanilla javascript. The data persists in local memory and the user can add and delete tasks as well as filter them by name.

This is just a small app I made along a Udemy course to refresh my memory on basic web development skills, DOM manipulation, vanilla javascript and html.